integration-tests/requirements.txt
# Integration Test Requirements
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0
requests>=2.28.0
zeep>=4.2.0  # SOAP client for ONVIF
aiohttp>=3.8.0  # For async HTTP requests
python-multipart>=0.0.5
xmltodict>=0.13.0
lxml>=4.9.0
python-dateutil>=2.8.0
urllib3>=1.26.0
pytest-html>=3.1.0  # For HTML reports
pytest-cov>=4.0.0  # For coverage reports
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
paramiko>=3.0.0  # For SSH device access (optional)
scapy>=2.4.0  # For network packet analysis
opencv-python>=4.7.0  # For video stream analysis
numpy>=1.24.0
Pillow>=9.0.0  # For image processing
psutil>=5.9.0  # For system monitoring during tests
python-dotenv>=1.0.0  # For loading environment variables from .env file
